-Fix: don't unlink(NULL)
git-svn-id: svn://repo.hu/pcb-rnd/trunk@2098 1d61de99-560f-4dba-8957-1e505d42358b

+++ src/plug_io.c
@@ -686,7 +686,8 @@ void SaveTMPData(void)
  */
 void RemoveTMPData(void)
 {
-	unlink(TMPFilename);
+	if (TMPFilename != NULL)
+		unlink(TMPFilename);
 }
 #endif
